knas

+++ routes/web.php
@@ -17,13 +17,13 @@ use Illuminate\Support\Facades\Route;
 //     return view('welcome');
 // });
 
-Route::get('/create', 'UserController@create')->name('posts.create');
+Route::get('/create', 'UserController@create')->name('posts.create')->middleware('auth');
 Route::post('/addpost', 'UserController@store')->name('addpost');
 Route::get('/', 'UserController@index');
 // Route::get('/userpost/{id}', 'UserController@show');
 
-Route::get('/posts/{id}', 'UserController@show')->name('posts.show');
-Route::delete('/posts/{id}', 'UserController@destroy')->name('posts.destroy');
+Route::get('/posts/{id}', 'UserController@show')->name('posts.show')->middleware('auth');
+Route::delete('/posts/{id}', 'UserController@destroy')->name('posts.destroy')->middleware('auth');
 Route::put('');
 Route::post('/comment/{id}', 'UserController@comment');
 
@@ -31,7 +31,12 @@ Route::get('/animals', 'PageController@animals')->name('animalpage');
 Route::get('/funnythings', 'PageController@funny')->name('funnypage');
 Route::get('/randomthings', 'PageController@random')->name('randompage');
 Route::get('/sport', 'PageController@sport')->name('sportpage');
+// Route::get('/user/{id}', 'UserController@shows');
 
 Auth::routes();
 
-Route::get('/home', 'HomeController@index')->name('home');
+Route::get('/home', 'HomeController@index')->name('home')->middleware('auth');
+
+Auth::routes();
+
+
